FEATURE: Read the description
This commit added a lot of different parts of code. First, main() was
fleshed out and now uses dotenv and env_logging, as well as implements a
timeout mechanism when serenity.rs keeps trying to connect. Second, made a
SlashCommand trait for all future commands to implement from. Third, added a
lot of logging statements.